% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_county_name.R
\name{find_countycode_fips}
\alias{find_countycode_fips}
\title{Find county code and fips}
\usage{
find_countycode_fips(county_to_find)
}
\arguments{
\item{county_to_find}{full name of county(ies)#'}
}
\value{
tibble with county code and fips
}
\description{
Find county code and fips
}
\examples{
find_countycode_fips("Dane")
}
