% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_county_name.R
\name{find_municipality_codes}
\alias{find_municipality_codes}
\title{Find municode by municipality}
\usage{
find_municipality_codes(municipality_to_find, ctv)
}
\arguments{
\item{municipality_to_find}{name of municipality (i.e. "Adams")}

\item{ctv}{c("City", "Town", "Village")}
}
\value{
df with municode and county code
}
\description{
Find municode by municipality
}
\examples{
data(muni)
find_municipality_codes("Adams", "Town")
}
