% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_groups.R
\name{get_age_groups}
\alias{get_age_groups}
\title{Bin age groups by 5 or 10 years (new db)}
\usage{
get_age_groups(person_df, bin_by = "5_yr")
}
\arguments{
\item{person_df}{person dataframe}

\item{bin_by}{select either "5_yr" or "10_yr"}
}
\value{
A new column called \emph{age_group_5yr} or \emph{age_group_10yr}
}
\description{
This bins ages into age groups by 5 years or 10 years.
}
\examples{
\dontrun{get_age_groups(person17, group_by = "10_yr"}
}
