% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crashflags_person.R
\name{get_alc_drug_impaired_person}
\alias{get_alc_drug_impaired_person}
\title{Get alcohol or drug person, or by driver}
\usage{
get_alc_drug_impaired_person(
  person_df,
  driver_only = "N",
  include_alc = "Y",
  include_drug = "Y",
  by = "and"
)
}
\arguments{
\item{person_df}{person df (new or old)}

\item{driver_only}{role as driver?}

\item{include_alc}{suspected alcohol?}

\item{include_drug}{suspected drug?}

\item{by}{intersection by - either "and" or "or"}
}
\value{
same person_df with drug_flag or alcohol_flag
}
\description{
Get alcohol or drug person, or by driver
}
