% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crashflags_crash.R
\name{get_crash_location}
\alias{get_crash_location}
\title{Get location of crash}
\usage{
get_crash_location(crash_df)
}
\arguments{
\item{crash_df}{crash dataframe w/ c("RLTNTRWY", "CRSHLOC", "INTTYPE", "INTDIS")

Find if crash is intersection, non-intersection, parking lot, or private property.}
}
\value{
A new column called \emph{crash_location}. This is for old and new bd, but
note that old db doesn't include parking lot/private property crashes.
}
\description{
Get location of crash
}
\examples{
\dontrun{get_crash_location(crash17)}
}
