% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crashflags_crash.R
\name{get_deerflag_crashes}
\alias{get_deerflag_crashes}
\title{Get flag for deer crashes (old and new db)}
\usage{
get_deerflag_crashes(crash_df)
}
\arguments{
\item{crash_df}{crash dataframe}
}
\value{
A new column called \emph{deer_flag} ("Y" or "N")
}
\description{
This finds if a crash involved a deer. Need \strong{CRSHTYPE} and \strong{ANMLTY}.
}
\examples{
\dontrun{get_deerflag_crashes(crash17)}
}
