% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crashflags_person.R
\name{get_distracted_driver_flag}
\alias{get_distracted_driver_flag}
\title{Get distracted driver flag}
\usage{
get_distracted_driver_flag(person_df)
}
\arguments{
\item{person_df}{data frame of person, must have ROLE, UNITNMBR, DISTACT,
DRVRDS}
}
\value{
speed_flag (Y or N)
}
\description{
This returns a column distracted == Y.
}
