% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_imports.R
\name{get_list_of_years}
\alias{get_list_of_years}
\title{Get a list of all years in between two dates}
\usage{
get_list_of_years(start_year, end_year)
}
\arguments{
\item{start_year}{start year of list, lowest is 1985.}

\item{end_year}{end year of list, highest is 2030.}
}
\value{
list of dates for years and years_old for data import
}
\description{
Useful when importing data among many years. Must be exactly two digits (i.e. 1992 would be "92").
}
\examples{
get_list_of_years("99", "14")
}
