% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_motorcyclists.R
\name{get_motorcycle_persons}
\alias{get_motorcycle_persons}
\title{Get motorcyclists (old and new db)}
\usage{
get_motorcycle_persons(person_df, vehicle_df)
}
\arguments{
\item{person_df}{person dataframe}

\item{vehicle_df}{vehicle dataframe}
}
\value{
Only motorcyclists in a crash
}
\description{
This finds if a person was a motorcyclist (driver or passenger).
}
\examples{
\dontrun{get_motorcycle_persons(person17, vehicle17)}
}
