% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_imports_narrative.R
\name{import_narrative}
\alias{import_narrative}
\title{Import crash narratives}
\usage{
import_narrative(filepath, years)
}
\arguments{
\item{filepath}{path where CSVs are stored (must all be in this folder)}

\item{years}{Year(s) of data c("20", "21"). Must be "17" or higher.}
}
\value{
crash narratives dataframe
}
\description{
Import crash narratives.
}
\examples{
import_narrative(filepath = "C:/CSV/csv_from_sas/from_sas_csv/", years = c("17","18"))
}
