% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{muni}
\alias{muni}
\title{lookup for municode}
\format{
rda with 5 columns

\describe{
\item{\code{Municipality}}{character COLUMN_DESCRIPTION}
\item{\code{Municipality_CTV}}{character COLUMN_DESCRIPTION}
\item{\code{MuniTownype}}{character COLUMN_DESCRIPTION}
\item{\code{MuniCode}}{double COLUMN_DESCRIPTION}
\item{\code{CntyCode}}{double COLUMN_DESCRIPTION}
}
}
\source{
wisdot
}
\usage{
muni
}
\description{
lookup for municode
}
\keyword{datasets}
