% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisdotcrashdatabase.R
\docType{package}
\name{wisdotcrashdatabase}
\alias{wisdotcrashdatabase}
\title{wisdotcrashdatabase: A package for doing crash data pulls and analysis}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

