% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_county_name.R
\name{county_rename}
\alias{county_rename}
\title{Get county name from CNTYCODE}
\usage{
county_rename(any_df)
}
\arguments{
\item{any_df}{person, crash, or vehicle dataframe}
}
\value{
Same dataframe, countyname column
}
\description{
CNTYCODE must be an integer.
}
\examples{
\dontrun{county_rename(vehicle17)}
}
