% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crash_times.R
\name{get_crash_times}
\alias{get_crash_times}
\title{Get crash hour}
\usage{
get_crash_times(dataframe)
}
\arguments{
\item{dataframe}{dataframe}
}
\value{
Adds a column 'newtime' with crash hour. i.e. "12am"
}
\description{
Get crash hour
}
\examples{
\dontrun{get_crash_times(crash17)}
}
