% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_imports.R
\name{get_db_data}
\alias{get_db_data}
\title{Puts data in a list of filenames to import}
\usage{
get_db_data(
  filepath,
  db_type,
  years_selected,
  years_selected_old,
  selected_columns = c("CRSHTIME")
)
}
\arguments{
\item{filepath}{path where CSVs are stored}

\item{db_type}{Type of database - any one of (crash, vehicle, person)}

\item{years_selected}{Year(s) of data c("20", "21")}

\item{years_selected_old}{Year(s) of data c("16")}

\item{selected_columns}{Columns to select, CRSHNMBR,CRSHDATE are automatically included}
}
\value{
data frame of db_type
}
\description{
it does this yada yada.
}
\examples{
get_db_data(filepath = "C:/CSV/csv_from_sas/fst/", db_type = "crash",
years_selected = c("17","18"), years_selected_old = c("15", "16"))
}
