% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crashflags_crash.R
\name{get_deerflag_crashes}
\alias{get_deerflag_crashes}
\title{Get deer flag for deer crashes}
\usage{
get_deerflag_crashes(crash_df)
}
\arguments{
\item{crash_df}{crash dataframe}
}
\value{
crash dataframe with deer_flag column ("Y" or "N")
}
\description{
Need CRSHTYPE and ANMLTY.
}
\examples{
\dontrun{get_deerflag_crashes(crash17)}
}
