% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crashflags_person.R
\name{get_driver_flags}
\alias{get_driver_flags}
\title{Get driver flags}
\usage{
get_driver_flags(person_df, flags)
}
\arguments{
\item{person_df}{data frame of person, must have DISTACT, DRVRDS}

\item{flags}{flags to select ("distracted","speed","teen","older")}
}
\value{
Adds a column of selected flags with distracted_flag ("Y" or "N")
}
\description{
This adds a column(s) of certain crash flags. It will only add a flag to
where there is a driver. Driver flags are: distracted, speed, teen, and
older.
}
\examples{
\dontrun{get_driver_flags(person_df, flags = c("teen", "distracted"))}

}
