% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crashflags_person.R
\name{relabel_person_variables}
\alias{relabel_person_variables}
\title{Relabels data in person}
\usage{
relabel_person_variables(person_df, relabel_by = "wisinj")
}
\arguments{
\item{person_df}{data frame of person, must have DISTACT, DRVRDS}

\item{relabel_by}{"wisinj", "bikeped"}
}
\value{
person_df with new column(s)
}
\description{
This bins certain variables by recategorizing and making a new column. This
is useful when working with data from an old and new database.
}
\examples{
\dontrun{system.file("extdata", "17person.fst", package = "fst") \%>\%
  relabel_person_date(relabel_by = "bikeped")}
}
